move ExpectedAssignmentValue to assignment package
